Rename `config/config` to `config/object`, remove `PackageConfig` mock
